* commit 'c733abfaccbe0403cc30fff9769611b2a03d346b':
  bug
  洗车流程bug

+++ XiaoMa/Class/HKCoupon.h
@@ -31,7 +31,7 @@ typedef enum : NSUInteger {
 @property (nonatomic,copy)NSString * couponName;
 
 ///数额
-@property (nonatomic)NSInteger couponAmount;
+@property (nonatomic)CGFloat couponAmount;
 
 ///优惠券描述
 @property (nonatomic,copy)NSString * couponDescription;
